Reorganize code following clean arch principles
