+++ app/src/main/java/com/vdemelo/starwarswiki/data/remote/NetworkConstants.kt
@@ -1,0 +1,7 @@
+package com.vdemelo.starwarswiki.data.remote
+
+object NetworkConstants {
+
+    const val BASE_URL = "https://swapi.dev/api/"
+
+}
