+++ app/src/main/java/com/vdemelo/starwarswiki/data/repository/StarWarsRemoteRepositoryImpl.kt
@@ -2,12 +2,13 @@ package com.vdemelo.starwarswiki.data.repository
 
 import com.vdemelo.starwarswiki.data.remote.api.StarWarsApi
 import com.vdemelo.starwarswiki.data.remote.response.SpeciesListResponse
+import com.vdemelo.starwarswiki.domain.repository.StarWarsRemoteRepository
 
-class StarWarsRepository(
+class StarWarsRemoteRepositoryImpl(
     private val api: StarWarsApi
-) {
+): StarWarsRemoteRepository {
 
-    suspend fun fetchSpecies(): SpeciesListResponse {
+    override suspend fun fetchSpecies(): SpeciesListResponse {
         return api.getSpecies()
     }
