+++ app/src/main/java/com/vdemelo/starwarswiki/domain/entity/model/SpeciesList.kt
@@ -1,16 +1,16 @@
-package com.vdemelo.starwarswiki.model
+package com.vdemelo.starwarswiki.domain.entity.model
 
-import com.vdemelo.starwarswiki.data.remote.response.PlanetsListResponse
+import com.vdemelo.starwarswiki.data.remote.response.SpeciesListResponse
 import com.vdemelo.starwarswiki.utils.nonNullOrEmpty
 
-class PlanetsList(
+class SpeciesList(
     val count: Int,
     val next: String?,
     val previous: String?,
-    val results: List<Planet>
+    val results: List<Species>
 ) {
 
-    constructor(response: PlanetsListResponse): this(
+    constructor(response: SpeciesListResponse): this(
         count = response.count ?: 0,
         next = response.next,
         previous = response.previous,
