+++ app/src/main/java/com/vdemelo/starwarswiki/domain/repository/StarWarsRemoteRepository.kt
@@ -1,0 +1,9 @@
+package com.vdemelo.starwarswiki.domain.repository
+
+import com.vdemelo.starwarswiki.data.remote.response.SpeciesListResponse
+
+interface StarWarsRemoteRepository {
+
+    suspend fun fetchSpecies(): SpeciesListResponse
+
+}
