+++ app/src/main/java/com/vdemelo/starwarswiki/domain/usecase/HomeUseCase.kt
@@ -1,0 +1,5 @@
+package com.vdemelo.starwarswiki.domain.usecase
+
+class HomeUseCase
+{
+}
