+++ app/src/main/java/com/vdemelo/starwarswiki/ui/HomeViewModel.kt
@@ -4,13 +4,13 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.vdemelo.starwarswiki.data.repository.StarWarsRepository
-import com.vdemelo.starwarswiki.model.Species
-import com.vdemelo.starwarswiki.model.SpeciesList
+import com.vdemelo.starwarswiki.data.repository.StarWarsRemoteRepositoryImpl
+import com.vdemelo.starwarswiki.domain.entity.model.Species
+import com.vdemelo.starwarswiki.domain.entity.model.SpeciesList
 import kotlinx.coroutines.launch
 
 class HomeViewModel(
-    private val repository: StarWarsRepository
+    private val repository: StarWarsRemoteRepositoryImpl
 ): ViewModel() {
 
     private val _species = MutableLiveData<Species?>()
