--- app/src/main/java/com/vdemelo/starwarswiki/utils/Constants.kt
@@ -1,7 +1,0 @@
-package com.vdemelo.starwarswiki.utils
-
-object Constants {
-
-    const val BASE_URL = "https://swapi.dev/api/"
-
-}
